Added state name and transition lines to editor

+++ src/editor/entity_editor/unpacked_state.h
@@ -14,6 +14,6 @@ struct UnpackedState {
     // Event, state
     std::map<std::string, std::string> next_states;
 
-    int x = 0;
-    int y = 0;
+    float x = 0.0;
+    float y = 0.0;
 };
